Changed from using separate methods for display images/files to just having one method for retrieving image objects by filename and another method for displaying image objects on a specified display.
displayWinner is now renderWinner which takes a image object, adds a
win message and returns it. If the Cubs win, the 'W' image is returned
instead.

showTeams is now displayLogos. To make future projects of adding
information to team logos easier, displayLogos takes kwargs of
alterations to be made to a team's logo and handles them before
displaying the images.