Did not realize that when a game is delayed, the elapsed_time var in the scoreboard object will say something like '2:55 (1:35 delay)' instead of just '2:55'.
So, an if branch has been introduced to determine whether the word
'delay' is in the elapsed time. If it is, some regex is used to extract
both times, then they are split at the colon (:), and converted to int
type. If any whole hours are in the minutes (>= 60 minutes), they are
moved to hours and minutes holds the remainder of minutes.

Also set that if elapsed_time is empty, meaning the game is likely
postponed, the game is treated as an old final.