Was having issues when a time string from the elapsed time attribute contained nothing on one side of the colon. For instance, if the game was delayed less than an hour, this would be the elapsed time: '2:18 (:36 delay)' (36 minutes).
So to fix this problem, an extra method has been implemented:
'hours_minutes_convert'. It takes a list of the hours and minutes in
string form and returns a tuple of hours and minutes in int form. If one
of the values is blank, it is treated as a 0.

This should prevent any errors related to the getFinals method for a
while.