Fixed issue that was causing segment displays to not turn off after all games were over.
When the update_cache method in the segment controller was passed a key
of 'off' it would turn off the segment displays, but then, near the
bottom of the method, it would immediately run either updateSingle or
updateDouble, thus turning the displays back on.

A new boolean attribute 'on' has been added that signifies whether the
displays are on or off. If the displays are off, the displays will not
be updated in the run method. When update_cache is run, the 'on'
attribute is initially set to True, but if the 'off' key is found in the
dict of changes, 'on' will be set to False. At the end, if the displays
are off, neither of the two update methods will be run.