Did tons of refactoring and organization of the different modules. Now, most of the Python files are in the cubbie_board directory. From there, the files are divided up into lcd_display, segment_display and core directory modules.
Also am working to change the way daemon execution is stopped. Rather
than checking if a file exists, the main module will have a catch
function that will shut everything down when a SIGTERM or SIGHUP signal
is received.

Also continued to add more content to the README. Should be complete
soon.